refactor(checker): split alias-pool and interface-depth guards into own modules
The previous commit's reset additions pushed two already-near-limit files over
the hard 2000-line ceiling (`type_alias_checking.rs` 2023, `cross_file.rs`
2018). Per the file-size rule the fix is to split, not ratchet.

- Extract the `ALIAS_DEFID_VISITED_POOL` thread-local, its `with_alias_defid_visited`
  accessor, reset, and test helpers into `alias_defid_visited_pool.rs`.
- Extract the `CROSS_ARENA_INTERFACE_DEPTH` thread-local, its enter/leave/in
  delegation methods, reset, and test helpers into `cross_file_interface_depth.rs`.

Both new modules are behavior-preserving moves; the aggregators and the
`clear_all_thread_local_state` wiring now route through them. `type_alias_checking.rs`
is 1970 lines and `cross_file.rs` is 1981 lines; the architecture guard passes.
