+++ crates/tsz-checker/src/state/type_analysis/cross_file.rs
@@ -14,44 +14,7 @@ use tsz_solver::TypeId;
 
 pub(crate) use super::cross_file_query_types::CrossFileQueryKind;
 
-thread_local! {
-    static CROSS_ARENA_INTERFACE_DEPTH: std::cell::Cell<u32> = const { std::cell::Cell::new(0) };
-}
-
-/// Reset the cross-arena interface-delegation depth counter to zero.
-///
-/// `enter_cross_arena_interface_delegation` / `leave_...` form a manual
-/// (non-RAII) pair, so an early bail-out between them (stack-overflow breaker,
-/// resolution fuel exhaustion) can leave the counter non-zero and suppress
-/// interface delegation in a later, unrelated compilation. Reset between
-/// independent compilations in batch mode.
-pub(crate) fn reset_cross_arena_interface_depth() {
-    CROSS_ARENA_INTERFACE_DEPTH.with(|c| c.set(0));
-}
-
-#[cfg(test)]
-pub(crate) fn set_cross_arena_interface_depth_for_test(value: u32) {
-    CROSS_ARENA_INTERFACE_DEPTH.with(|c| c.set(value));
-}
-
-#[cfg(test)]
-pub(crate) fn cross_arena_interface_depth_for_test() -> u32 {
-    CROSS_ARENA_INTERFACE_DEPTH.with(std::cell::Cell::get)
-}
-
 impl<'a> CheckerState<'a> {
-    pub(crate) fn enter_cross_arena_interface_delegation() {
-        CROSS_ARENA_INTERFACE_DEPTH.with(|c| c.set(c.get() + 1));
-    }
-
-    pub(crate) fn leave_cross_arena_interface_delegation() {
-        CROSS_ARENA_INTERFACE_DEPTH.with(|c| c.set(c.get().saturating_sub(1)));
-    }
-
-    pub(crate) fn in_cross_arena_interface_delegation() -> bool {
-        CROSS_ARENA_INTERFACE_DEPTH.with(|c| c.get() > 0)
-    }
-
     fn resolve_cross_file_heritage_type_arg(
         &mut self,
         arena: &tsz_parser::NodeArena,
