+++ crates/tsz-checker/src/state/type_analysis/mod.rs
@@ -26,6 +26,7 @@ mod cross_file_direct_functions;
 mod cross_file_env_merge;
 mod cross_file_globals;
 mod cross_file_import_alias_pin;
+mod cross_file_interface_depth;
 mod cross_file_overlay_gate;
 pub(crate) mod cross_file_query_types;
 mod cross_file_residue;
@@ -42,18 +43,18 @@ mod type_param_defaults;
 /// leave a non-zero depth counter or a non-empty resolution stack that would
 /// suppress cross-file resolution in the next project on the same worker.
 pub(crate) fn reset_cross_file_recursion_guards() {
-    cross_file::reset_cross_arena_interface_depth();
+    cross_file_interface_depth::reset_cross_arena_interface_depth();
     cross_file_alias_cycle::reset_cross_arena_alias_stack();
 }
 
 #[cfg(test)]
 pub(crate) fn dirty_cross_file_recursion_guards_for_test() {
-    cross_file::set_cross_arena_interface_depth_for_test(2);
+    cross_file_interface_depth::set_cross_arena_interface_depth_for_test(2);
     cross_file_alias_cycle::push_cross_arena_alias_for_test(tsz_solver::def::DefId::INVALID);
 }
 
 #[cfg(test)]
 pub(crate) fn cross_file_recursion_guards_clear_for_test() -> bool {
-    cross_file::cross_arena_interface_depth_for_test() == 0
+    cross_file_interface_depth::cross_arena_interface_depth_for_test() == 0
         && cross_file_alias_cycle::cross_arena_alias_stack_len_for_test() == 0
 }
